Make get_disease a top level function in dtda.py
