+++ bioagents/dtda/dtda.py
@@ -7,6 +7,7 @@ import re
 import os
 import numpy
 import logging
+import xml.etree.ElementTree as ET
 
 from indra.sources.indra_db_rest import get_statements
 from indra.databases import cbio_client, hgnc_client
@@ -302,6 +303,24 @@ class Disease(object):
 
     def __str__(self):
         return self.__repr__()
+
+
+def get_disease(disease_str):
+        term = ET.fromstring(disease_str).find('TERM')
+        disease_type = term.find('type').text
+        if disease_type.startswith('ONT::'):
+            disease_type = disease_type[5:].lower()
+        drum_term = term.find('drum-terms/drum-term')
+        if drum_term is None:
+            dbname = term.find('name').text
+            dbid_dict = {}
+        else:
+            dbname = drum_term.attrib['name']
+            dbid = term.attrib['dbid']
+            dbids = dbid.split('|')
+            dbid_dict = {k: v for k, v in [d.split(':') for d in dbids]}
+        disease = Disease(disease_type, dbname, dbid_dict)
+        return disease
 
 
 def _convert_term(term):
